+++ examples/frequently-bought-together/fbt-streaming-dtc.yml
@@ -5,7 +5,7 @@
 # Type of DTC. Streaming or Window
 # The Streaming DTC processes the raw data and converts it into sessions
 
-Type: Blurr:Streaming
+Type: Blurr:Transform:Streaming
 Version: '2018-03-01'
 Description: Level 1 transformation from streaming Raw Data
 Name: fbt_v1
