+++ examples/offer-ai/offer-ai-window-dtc.yml
@@ -4,7 +4,7 @@
 
 # Type of DTC. Streaming or Window
 # The Window DTC periodically applies windows operations on top of sessionized data
-Type: Blurr:Window
+Type: Blurr:Transform:Window
 Version: '2018-03-01'
 Description: Window ops for feature generation
