+++ tests/core/expression_test.py
@@ -9,14 +9,14 @@ from blurr.core.base import Expression
 from blurr.core.errors import InvalidExpressionError, MissingAttributeError
 from blurr.core.evaluation import Context, EvaluationContext
 from blurr.core.schema_loader import SchemaLoader
-from blurr.core.streaming_transformer import StreamingTransformer
+from blurr.core.transformer_streaming import StreamingTransformer
 
 
 @fixture
 def schema_spec() -> Dict:
     return {
         'Name': 'test',
-        'Type': 'Blurr:Transform:Simple',
+        'Type': 'Blurr:Transform:Streaming',
         "Version": "2018-03-01",
         "Time": "parser.parse(source.event_time)",
         "Identity": "source.user_id",
