+++ tests/core/streaming_transformer_test.py
@@ -7,14 +7,14 @@ from blurr.core.errors import IdentityError, TimeError
 from blurr.core.evaluation import Context
 from blurr.core.record import Record
 from blurr.core.schema_loader import SchemaLoader
-from blurr.core.streaming_transformer import StreamingTransformerSchema, StreamingTransformer
+from blurr.core.transformer_streaming import StreamingTransformerSchema, StreamingTransformer
 
 
 @fixture
 def schema_spec() -> Dict[str, Any]:
     return {
         'Name': 'test',
-        'Type': 'Blurr:Streaming',
+        'Type': 'Blurr:Transform:Streaming',
         'Version': '2018-03-01',
         'Identity': '\'user1\'',
         'Time': 'datetime(2016,10,10)',
