+++ tests/core/syntax/dtcs/invalid_datagroup_has_no_fields.yml
@@ -1,4 +1,4 @@
-Type: Blurr:Streaming
+Type: Blurr:Transform:Streaming
 Version: '2018-03-01'
 Description: Example
 Name: example_name
