+++ tests/core/syntax/dtcs/invalid_incorrect_expression.yml
@@ -1,4 +1,4 @@
-Type: Blurr:Streaming
+Type: Blurr:Transform:Streaming
 Version: '2018-03-01'
 Description: Example
 Name: example_name
