+++ tests/core/syntax/dtcs/invalid_string_instead_integer.yml
@@ -1,4 +1,4 @@
-Type: Blurr:Window
+Type: Blurr:Transform:Window
 Version: '2018-03-01'
 Description: Example
 SourceDTC: test
