+++ tests/core/syntax/dtcs/invalid_wrong_version.yml
@@ -1,4 +1,4 @@
-Type: Blurr:Streaming
+Type: Blurr:Transform:Streaming
 
 
 # WRONG VERSION
