+++ tests/core/syntax/dtcs/invalid_yaml.yml
@@ -1,4 +1,4 @@
-Type: Blurr:Streaming
+Type: Blurr:Transform:Streaming
 
 
       Description: Example
