+++ tests/core/window_transformer_test.py
@@ -10,8 +10,8 @@ from blurr.core.evaluation import Context, EvaluationContext
 from blurr.core.schema_loader import SchemaLoader
 from blurr.core.aggregate_block import BlockAggregate, \
     BlockAggregateSchema
-from blurr.core.streaming_transformer import StreamingTransformer
-from blurr.core.window_transformer import WindowTransformer, \
+from blurr.core.transformer_streaming import StreamingTransformer
+from blurr.core.transformer_window import WindowTransformer, \
     WindowTransformerSchema
 from tests.core.conftest import init_memory_store
